+++ QuizMaster.API.Services/MockQuestionService.cs
@@ -1,6 +1,5 @@
 using System.Collections.Generic;
 using System.Linq;
-using QuizMaster.API.Data;
 using QuizMaster.API.Services.Interfaces;
 using QuizMaster.Shared.Models;
 
@@ -33,14 +32,6 @@ namespace QuizMaster.API.Services
                     Answers = answers.Where(x => x.Id == 8 || x.Id == 9 || x.Id == 10)
                 }
             };
-        }
-    }
-
-    public class QuestionService : DataService<Question>
-    {
-        public QuestionService(QuizContext context) : base(context)
-        {
-            
         }
     }
 }
