+++ QuizMaster.API.Services/MockQuizService.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using QuizMaster.API.Data;
 using QuizMaster.API.Services.Interfaces;
 using QuizMaster.Shared.Models;
 
@@ -13,13 +12,6 @@ namespace QuizMaster.API.Services
             {
                 new Quiz {Id = 0, Title = "Test Quiz", Rounds = roundService.GetAllAsync().Result}
             };
-        }
-    }
-
-    public class QuizService : DataService<Quiz>
-    {
-        public QuizService(QuizContext context) : base(context)
-        {
         }
     }
 }
