+++ QuizMaster.API.Services/QuestionService.cs
@@ -1,0 +1,12 @@
+using QuizMaster.API.Data;
+using QuizMaster.Shared.Models;
+
+namespace QuizMaster.API.Services
+{
+    public class QuestionService : DataService<Question>
+    {
+        public QuestionService(QuizContext context) : base(context)
+        {
+        }
+    }
+}
