+++ QuizMaster.API.Services/RoundService.cs
@@ -1,0 +1,12 @@
+using QuizMaster.API.Data;
+using QuizMaster.Shared.Models;
+
+namespace QuizMaster.API.Services
+{
+    public class RoundService : DataService<Round>
+    {
+        public RoundService(QuizContext context) : base(context)
+        {
+        }
+    }
+}
